Move `component.js` to component-specific folder
