+++ packages/components/src/accessible-popover/index.js
@@ -1,2 +1,2 @@
-export { default as AccessiblePopover } from './component';
+export { default as AccessiblePopover } from './accessible-popover';
 export * from './context';
